+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-05-29 22:10:08
-// Commit: test(i18n): implement widget integration tests for i18n
+// Updated at: 2026-05-30 09:56:48
+// Commit: feat(i18n): integrate custom telemetry logger in i18n
 
 class i18nComponentState {
-  final String commitRef = '1602';
-  final String lastMessage = 'test(i18n): implement widget integration tests for i18n';
-  final DateTime timestamp = DateTime.parse('2026-05-29T22:10:08+05:00');
+  final String commitRef = '1607';
+  final String lastMessage = 'feat(i18n): integrate custom telemetry logger in i18n';
+  final DateTime timestamp = DateTime.parse('2026-05-30T09:56:48+05:00');
 }