+++ CHANGELOG.md
@@ -1608,3 +1608,4 @@
 - [2026-05-30 09:07] perf(core): apply RepaintBoundary on animated core widgets
 - [2026-05-30 09:12] feat(navigation): add bi-directional event bus handler for navigation
 - [2026-05-30 09:39] style(theme): harmonize light and dark theme palette for theme
+- [2026-05-30 09:56] feat(i18n): integrate custom telemetry logger in i18n